Update AllModels.Rmd
Added mclust and fixed up neural nets. For loop was removed as the stepmax needed to just be increased by a substantial amount, but nonetheless did not rove totally useful as we have already found from practice that Pr_Male runs best with 3 nodes. mclust was not particularly helpful, but added a discussion as to why within the document as well for those that want to look over it :)